+++ lib/printer.mli
@@ -13,7 +13,7 @@ val make_debug_format : int -> string -> bool -> string -> string
     containing these parameters. *)
 
 val default_cost_factory : page_width:int -> ?computation_width:int -> unit ->
-                           (module Signature.CostFactory with type t = int * int * int * int)
+                           (module Signature.CostFactory with type t = int * int * int)
 (** The default cost factory, parameterized by the page width limit [page_width],
     and optionally {{!page-index.complimit}the computation width limit}
     [computation_width].
@@ -25,16 +25,14 @@ val default_cost_factory : page_width:int -> ?computation_width:int -> unit ->
     {ul {- The first component is {i badness}, which is roughly speaking
            the sum of squared overflows over the page width limit}
         {- The second component is sum of overflows over a column separator.}
-        {- The third component is the height (number of newlines).}
-        {- The fourth component is bias penalty to encourage toward choosing
-           a leftmost column separator.} }
+        {- The third component is the height (number of newlines).}}
 
     Internally, [default_cost_factory] is defined as:
 
     {@ocaml file=printer.ml,part=default_cost_factory[
 let default_cost_factory ~page_width ?computation_width () =
   (module struct
-    type t = int * int * int * int
+    type t = int * int * int
 
     let limit = match computation_width with
       | None -> (float_of_int page_width) *. 1.2 |> int_of_float
@@ -46,23 +44,26 @@ let default_cost_factory ~page_width ?computation_width () =
         let maxwc = max page_width pos in
         let a = maxwc - page_width in
         let b = stop - maxwc in
-        (b * (2*a + b), 0, 0, 0)
+        (b * (2*a + b), 0, 0)
       else
-        (0, 0, 0, 0)
+        (0, 0, 0)
 
-    let newline _ = (0, 0, 1, 0)
+    let newline _ = (0, 0, 1)
 
-    let combine (o1, ot1, h1, bt1) (o2, ot2, h2, bt2) =
-      (o1 + o2, ot1 + ot2, h1 + h2, bt1 + bt2)
+    let combine (o1, ot1, h1) (o2, ot2, h2) =
+      (o1 + o2, ot1 + ot2, h1 + h2)
 
     let le c1 c2 = c1 <= c2
 
-    let two_columns_overflow w = (0, w, 0, 0)
+    let two_columns_overflow w = (0, w, 0)
 
-    let two_columns_bias     w = (0, 0, 0, w)
+    let two_columns_bias     _ = (0, 0, 0)
 
-    let string_of_cost (o, ot, h, bt) = Printf.sprintf "(%d %d %d %d)" o ot h bt
+    let string_of_cost (o, ot, h) = Printf.sprintf "(%d %d %d)" o ot h
 
     let debug_format = make_debug_format page_width
-  end: Signature.CostFactory with type t = int * int * int * int)
+  end: Signature.CostFactory with type t = int * int * int)
 ]} *)
+
+val version : string
+(* a version string *)