+++ test/pretty_expressive.ml
@@ -123,15 +123,248 @@ let test_sexp_10 () =
        [ "(a b c d)" ])
     (print_sexp example_sexp 10)
 
+let test_two_columns_case_1 () =
+  let cf = Printer.default_cost_factory ~page_width:4 ~computation_width:100 () in
+  let module P = Printer.Make (val cf) in
+  let open P in
+
+  let d1 = text "x" <$>
+           text "xxx" in
+  let d2 = text "xxx" <$>
+           text "x" in
+  let d_below = text "" in
+  let d_right1 = text "zz" in
+  let d_right2 = text "wwww" in
+
+  Alcotest.(check string) "same string"
+    (String.concat "\n"
+       [ "1234";
+         "xxx │" ;
+         "xzz │";
+         "wwww│";
+         "";
+         "is_tainted: false";
+         "cost: (0 1 2)" ])
+    (pretty_format_debug (two_columns [ ( d1 <|> d2, d_right1 ) ;
+                                        ( d_below,   d_right2 ) ]))
+
+let test_two_columns_case_2 () =
+  let cf = Printer.default_cost_factory ~page_width:5 ~computation_width:100 () in
+  let module P = Printer.Make (val cf) in
+  let open P in
+
+  let d1 = text "x" <$>
+           text "xxx" in
+  let d2 = text "xxx" <$>
+           text "x" in
+  let d_below = text "qq" in
+  let d_right1 = text "zzz" in
+  let d_right2 = text "www" in
+
+  Alcotest.(check string) "same string"
+    (String.concat "\n"
+       [ "12345";
+         "xxx  │" ;
+         "x zzz│";
+         "qqwww│";
+         "";
+         "is_tainted: false";
+         "cost: (0 0 2)" ])
+    (pretty_format_debug (two_columns [ ( d1 <|> d2, d_right1 ) ;
+                                        ( d_below,   d_right2 ) ]))
+
+let test_two_columns_case_3 () =
+  let cf = Printer.default_cost_factory ~page_width:7 ~computation_width:100 () in
+  let module P = Printer.Make (val cf) in
+  let open P in
+
+  let d1 = text "x" <$>
+           text "xxx" in
+  let d2 = text "xxx" <$>
+           text "x" in
+  let d_below = text "qqqq" in
+  let d_right1 = text "zzz" in
+  let d_right2 = text "www" in
+
+  Alcotest.(check string) "same string"
+    (String.concat "\n"
+       [ "1234567";
+         "xxx    │" ;
+         "x   zzz│";
+         "qqqqwww│";
+         "";
+         "is_tainted: false";
+         "cost: (0 0 2)" ])
+    (pretty_format_debug (two_columns [ ( d1 <|> d2, d_right1 ) ;
+                                        ( d_below,   d_right2 ) ]))
+
+let test_two_columns_regression_phantom () =
+  let cf = Printer.default_cost_factory ~page_width:7 ~computation_width:100 () in
+  let module P = Printer.Make (val cf) in
+  let open P in
+
+  (* creates column separator at 3 *)
+  let phantom_doc = text "aaaaaaaaaaaaaaaa" <$>
+                    text "a" <$>
+                    text "aaa" in
+  let d = text "a" <$>
+          text "aaaa" in
+  let d_below = text "q" in
+  let d_right1 = text "zzz" in
+  let d_right2 = text "wwww" in
+
+  Alcotest.(check string) "same string"
+    (String.concat "\n"
+       [ "1234567";
+         "a      │" ;
+         "aaaazzz│";
+         "qwwww  │";
+         "";
+         "is_tainted: false";
+         "cost: (0 3 2)" ])
+    (pretty_format_debug (two_columns [ ( phantom_doc <|> d, d_right1 ) ;
+                                        ( d_below, d_right2 ) ]))
+
+(* This is a cost factory that cares more about preserving the two-column shape
+   than avoiding overflows *)
+let strict_two_columns_cost_factory ~page_width ?computation_width () =
+  (module struct
+    type t = int * int * int
+
+    let limit = match computation_width with
+      | None -> (float_of_int page_width) *. 1.2 |> int_of_float
+      | Some computation_width -> computation_width
+
+    let text pos len =
+      let stop = pos + len in
+      if stop > page_width then
+        let maxwc = max page_width pos in
+        let a = maxwc - page_width in
+        let b = stop - maxwc in
+        (0, b * (2*a + b), 0)
+      else
+        (0, 0, 0)
+
+    let newline _ = (0, 0, 1)
+
+    let combine (ot1, o1, h1) (ot2, o2, h2) =
+      (ot1 + ot2, o1 + o2, h1 + h2)
+
+    let le c1 c2 = c1 <= c2
+
+    let two_columns_overflow w = (w, 0, 0)
+
+    let two_columns_bias _ = (0, 0, 0)
+
+    let string_of_cost (ot, o, h) = Printf.sprintf "(%d %d %d)" ot o h
+
+    let debug_format = Printer.make_debug_format page_width
+  end: Signature.CostFactory with type t = int * int * int)
+
+let test_two_columns_factory_overflow () =
+  let cf = strict_two_columns_cost_factory ~page_width:4 ~computation_width:100 () in
+  let module P = Printer.Make (val cf) in
+  let open P in
+
+  let d1 = text "xxx" in
+  let d2 = text "xxxxx" <$>
+           text "x" in
+  let d_below = text "" in
+  let d_right1 = text "zz" in
+  let d_right2 = text "wwww" in
+
+  (* NOTE: choosing d2 is better than choosing d1,
+     since d_right1 will overflow as much as d2,
+     and d_right2 will overflow A LOT *)
+  Alcotest.(check string) "same string"
+    (String.concat "\n"
+       [ "1234";
+         "xxxx│x";
+         "xzz │";
+         " www│w";
+         "";
+         "is_tainted: false";
+         "cost: (0 2 2)" ])
+    (pretty_format_debug (two_columns [ ( d1 <|> d2, d_right1 ) ;
+                                        ( d_below,   d_right2 ) ]))
+
+(* This is a cost factory that cares about choosing leftmost separator, more than
+   minimizing number of lines. It still tries to avoid overflows though *)
+let biased_two_columns_cost_factory ~page_width ?computation_width () =
+  (module struct
+    type t = int * int * int * int
+
+    let limit = match computation_width with
+      | None -> (float_of_int page_width) *. 1.2 |> int_of_float
+      | Some computation_width -> computation_width
+
+    let text pos len =
+      let stop = pos + len in
+      if stop > page_width then
+        let maxwc = max page_width pos in
+        let a = maxwc - page_width in
+        let b = stop - maxwc in
+        (b * (2*a + b), 0, 0, 0)
+      else
+        (0, 0, 0, 0)
+
+    let newline _ = (0, 0, 0, 1)
+
+    let combine (o1, ot1, b1, h1) (o2, ot2, b2, h2) =
+      (o1 + o2, ot1 + ot2, b1 + b2, h1 + h2)
+
+    let le c1 c2 = c1 <= c2
+
+    let two_columns_overflow w = (0, w, 0, 0)
+
+    let two_columns_bias w = (0, 0, w, 0)
+
+    let string_of_cost (ot, o, b, h) = Printf.sprintf "(%d %d %d %d)" ot o b h
+
+    let debug_format = Printer.make_debug_format page_width
+  end: Signature.CostFactory with type t = int * int * int * int)
+
+let test_two_columns_factory_bias () =
+  let cf = biased_two_columns_cost_factory ~page_width:4 ~computation_width:100 () in
+  let module P = Printer.Make (val cf) in
+  let open P in
+
+  let d1 = text "xxxxx" <$>
+           text "x" <$>
+           text "x" in
+  let d2 = text "xx" <$>
+           text "xx" in
+  let d3 = text "xxx" in
+  let d_below = text "" in
+  let d_right1 = text "z" in
+  let d_right2 = text "w" in
+
+  Alcotest.(check string) "same string"
+    (String.concat "\n"
+       [ "1234";
+         "xx  │" ;
+         "xxz │";
+         "  w │";
+         "";
+         "is_tainted: false";
+         "cost: (0 0 2 2)" ])
+    (pretty_format_debug (two_columns [ ( d1 <|> d2 <|> d3, d_right1 ) ;
+                                        ( d_below,   d_right2 ) ]))
+
 let suite =
-  [ "choice; w = 80", `Quick, test_choice_doc_80 ;
-    "choice; w = 20", `Quick, test_choice_doc_20 ;
-    "group; w = 80", `Quick, test_group_doc_80 ;
-    "group; w = 20", `Quick, test_group_doc_20 ;
-    "sexp; w = 4", `Quick, test_sexp_4 ;
-    "sexp; w = 6", `Quick, test_sexp_6 ;
-    "sexp; w = 10", `Quick, test_sexp_10
-  ]
+  [ "choice; w = 80", `Quick, test_choice_doc_80;
+    "choice; w = 20", `Quick, test_choice_doc_20;
+    "group; w = 80", `Quick, test_group_doc_80;
+    "group; w = 20", `Quick, test_group_doc_20;
+    "sexp; w = 4", `Quick, test_sexp_4;
+    "sexp; w = 6", `Quick, test_sexp_6;
+    "sexp; w = 10", `Quick, test_sexp_10;
+    "two_columns (1)", `Quick, test_two_columns_case_1;
+    "two_columns (2)", `Quick, test_two_columns_case_2;
+    "two_columns (3)", `Quick, test_two_columns_case_3;
+    "two_columns (regression phantom space)", `Quick, test_two_columns_regression_phantom;
+    "two_columns (cost factory - overflow)", `Quick, test_two_columns_factory_overflow;
+    "two_columns (cost factory - bias)", `Quick, test_two_columns_factory_bias ]
 
 let () =
   Alcotest.run "pretty expressive" [ "example doc", suite ]